requirements.txt
numpy~=1.23.2
pandas~=1.4.3
python-dotenv~=0.20.0
aiogram~=2.22.1